refactor(SchemaAPI): move `SchemaAPI` into its own file
