+++ src/index.js
@@ -1,9 +1,10 @@
 import fetch from 'isomorphic-fetch';
 import querystring from 'querystring';
 import _ from 'lodash';
-import { SchemaAPI, UserAPI, RoleAPI } from './api';
+import { UserAPI, RoleAPI } from './api';
 import DocumentsAPI from './api/documents';
 import FileAPI from './api/file';
+import SchemaAPI from './api/schema';
 
 /**
  * Count bytes in a string's UTF-8 representation.
